chanstate: finish kv store extraction
Move the remaining channel-state tests and helpers into the chanstate
package so the store owns the KV-specific behavior directly.

This also narrows the channel-state store interface by removing
cross-domain link-node lifecycle methods. Those workflows will be owned
by a separate channel coordinator in the next commits.
