+++ chanstate/kv_close_summary.go
@@ -23,6 +23,12 @@ func ClosedChannelBucketKey() []byte {
 	return closedChannelBucket
 }
 
+// ClosedChannelBucket returns the top-level closed-channel summary bucket,
+// creating it if needed.
+func ClosedChannelBucket(tx kvdb.RwTx) (kvdb.RwBucket, error) {
+	return tx.CreateTopLevelBucket(closedChannelBucket)
+}
+
 // PutChannelCloseSummary writes the immutable close-time summary of a channel
 // under the closed channel bucket.
 func PutChannelCloseSummary(tx kvdb.RwTx, chanID []byte,
