+++ chanstate/kv_forwarding_package.go
@@ -85,24 +85,6 @@ func FwdPackagesBucketKey() []byte {
 	return fwdPackagesKey
 }
 
-// Encode serializes the AddRef to the given io.Writer.
-func (a *AddRef) Encode(w io.Writer) error {
-	if err := binary.Write(w, binary.BigEndian, a.Height); err != nil {
-		return err
-	}
-
-	return binary.Write(w, binary.BigEndian, a.Index)
-}
-
-// Decode deserializes the AddRef from the given io.Reader.
-func (a *AddRef) Decode(r io.Reader) error {
-	if err := binary.Read(r, binary.BigEndian, &a.Height); err != nil {
-		return err
-	}
-
-	return binary.Read(r, binary.BigEndian, &a.Index)
-}
-
 // Size returns number of bytes produced when the PkgFilter is serialized.
 func (f *PkgFilter) Size() uint16 {
 	// 2 bytes for uint16 `count`, then round up number of bytes required to
